fix issue with constrained targets in ion raster overlay ui
