+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes/__init__.py
@@ -1,7 +1,6 @@
 from .data_attributes_widget import CesiumDataSchemaAttributesWidget  # noqa: F401
 from .georeference_attributes_widget import CesiumGeoreferenceSchemaAttributesWidget  # noqa :F401
 from .tileset_attributes_widget import CesiumTilesetAttributesWidget  # noqa: F401
-from .raster_overlay_attributes_widget import CesiumRasterOverlayAttributesWidget  # noqa: F401
 from .globe_anchor_attributes_widget import CesiumGlobeAnchorAttributesWidget  # noqa: F401
 from .ion_server_attributes_widget import CesiumIonServerAttributesWidget  # noqa: F401
 from .ion_raster_overlay_attributes_widget import CesiumIonRasterOverlayAttributesWidget  # noqa: F401
