+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes/ion_raster_overlay_attributes_widget.py
@@ -3,7 +3,10 @@ import omni.ui as ui
 from typing import List
 from omni.kit.property.usd.custom_layout_helper import CustomLayoutFrame, CustomLayoutGroup, CustomLayoutProperty
 from omni.kit.property.usd.usd_property_widget import SchemaPropertiesWidget
-from cesium.usd.plugins.CesiumUsdSchemas import IonRasterOverlay as CesiumIonRasterOverlay
+from cesium.usd.plugins.CesiumUsdSchemas import (
+    IonRasterOverlay as CesiumIonRasterOverlay,
+    IonServer as CesiumIonServer,
+)
 from pxr import Sdf
 
 
@@ -84,3 +87,14 @@ class CesiumIonRasterOverlayAttributesWidget(SchemaPropertiesWidget):
                 CustomLayoutProperty("cesium:showCreditsOnScreen")
 
         return frame.apply(props)
+
+    def _filter_props_to_build(self, props):
+        filtered_props = super()._filter_props_to_build(props)
+        filtered_props.extend(prop for prop in props if prop.GetName() == "cesium:ionServerBinding")
+        return filtered_props
+
+    def get_additional_kwargs(self, ui_attr):
+        if ui_attr.prop_name == "cesium:ionServerBinding":
+            return None, {"target_picker_filter_type_list": [CesiumIonServer], "targets_limit": 1}
+
+        return None, {"targets_limit": 0}
