+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes_widget_controller.py
@@ -3,7 +3,6 @@ import omni.kit.window.property
 from .attributes import (
     CesiumDataSchemaAttributesWidget,
     CesiumGeoreferenceSchemaAttributesWidget,
-    CesiumRasterOverlayAttributesWidget,
     CesiumTilesetAttributesWidget,
     CesiumGlobeAnchorAttributesWidget,
     CesiumIonServerAttributesWidget,
